Changed some more things
For one thing, you can now go to the start of the line or to the end of the line with `<A-a>` or `<A-e>`, like you can in Emacs with `<C-a>` or `<C-e>`.

There's also `<A-w>` and `<A-b>` for moving word-wise.

Additionally, if you enter a key combination that isn't recognised, it'll be printed to the Echo Area, for example, "`<C-b>` is undefined."

Also, the Echo Area will now be cleared after three seconds from updating it. This is more of a visual thing and is functionally not much different.